refactor: relocate useQueryWorkflowTemplate to a more appropriate module and remove obsolete implementation
